sol comments
