+++ src/Components/PeopleMenu.tsx
@@ -1,28 +1,57 @@
-import React, { useCallback } from 'react';
+import React, { useMemo, useCallback, useState } from 'react';
 import { Person } from '../types/Person';
 import classNames from 'classnames';
 
 type Props = {
+  delay: number;
   people: Person[];
-  query: string;
-  isOpen: boolean;
-  setFocused: (focuse: boolean) => void;
-  setIsOpen: (isOpen: boolean) => void;
-  setQuery: (query: string) => void;
   setSelected: (person: Person | null) => void;
-  applyQuery: (setquery: string) => void;
 };
+
+function debounce<T extends unknown[]>(
+  callback: (...args: T) => void,
+  delay: number,
+) {
+  let timerId = 0;
+
+  return (...args: T) => {
+    window.clearTimeout(timerId);
+    timerId = window.setTimeout(() => {
+      callback(...args);
+    }, delay);
+  };
+}
 
 const PeopleMenuComponent: React.FC<Props> = ({
   people,
-  query,
-  isOpen,
-  setIsOpen,
-  setQuery,
-  setFocused,
   setSelected,
-  applyQuery,
+  delay,
 }) => {
+  const [isOpen, setIsOpen] = useState(false);
+  const [isFocused, setIsFocused] = useState(false);
+  const [query, setQuery] = useState('');
+  const [appliedQuery, setAppliedQuery] = useState('');
+
+  const applyQuery = useCallback(debounce(setAppliedQuery, delay), []);
+
+  const filterPeople = useMemo(() => {
+    if (!appliedQuery.trim()) {
+      return people;
+    }
+
+    return people.filter(person =>
+      person.name
+        .toLocaleLowerCase()
+        .includes(appliedQuery.toLocaleLowerCase()),
+    );
+  }, [appliedQuery]);
+
+  const visiblePeople = useMemo(() => {
+    return isFocused ? filterPeople : [];
+  }, [isFocused, filterPeople]);
+
+  const error = isFocused && query && filterPeople.length === 0;
+
   const handleInput = useCallback(
     (event: React.ChangeEvent<HTMLInputElement>) => {
       setQuery(event.target.value);
@@ -52,14 +81,20 @@ const PeopleMenuComponent: React.FC<Props> = ({
           className="input"
           data-cy="search-input"
           onChange={handleInput}
-          onFocus={() => setFocused(true)}
-          onBlur={() => setFocused(false)}
+          onFocus={() => {
+            setIsFocused(true);
+            setIsOpen(true);
+          }}
+          onBlur={() => {
+            setIsFocused(false);
+            setIsOpen(false);
+          }}
         />
       </div>
 
       <div className="dropdown-menu" role="menu" data-cy="suggestions-list">
         <div className="dropdown-content">
-          {people.map(person => (
+          {visiblePeople.map(person => (
             <a
               key={person.name}
               className="dropdown-item"
@@ -72,6 +107,21 @@ const PeopleMenuComponent: React.FC<Props> = ({
           ))}
         </div>
       </div>
+      {error && (
+        <div
+          className="
+            notification
+            is-danger
+            is-light
+            mt-3
+            is-align-self-flex-start
+          "
+          role="alert"
+          data-cy="no-suggestions-message"
+        >
+          <p className="has-text-danger">No matching suggestions</p>
+        </div>
+      )}
     </div>
   );
 };
